backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client==1.7.0
openai==1.10.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-openai==0.0.5
tiktoken==0.5.2
